feat: CLIN-2267 update
